Moved List outside of core. Fixes #336
