+++ src/Pair/toPairs.js
@@ -1,7 +1,6 @@
 /** @license ISC License (c) copyright 2017 original and current authors */
 /** @author Ian Hofmann-Hicks (evil) */
 
-const List = require('../core/List')
 const Pair = require('../core/Pair')
 const isObject = require('../core/isObject')
 
@@ -13,9 +12,9 @@ function toPairs(obj) {
 
   return Object.keys(obj).reduce(
     (acc, key) => obj[key] !== undefined
-      ? acc.concat(List.of(Pair(key, obj[key])))
+      ? acc.concat( [ Pair(key, obj[key]) ] )
       : acc,
-    List.empty()
+    []
   )
 }
